set 'png' compression format for depth images
By default the compressed_image_transport is using the jpeg format. This is
a lossy compression format which introduces compression artefacts which are
not visible in colour images, but corrupt the 16bit depth images.
Change the depth compression to the lossless png format to keep the original
data when decompressing depth images.